bugfix FullGeeftDeatails : don't cast in long,use double

+++ app/src/main/java/samurai/geeft/android/geeft/fragments/FullGeeftDeatailsFragment.java
@@ -614,7 +614,7 @@ public class FullGeeftDeatailsFragment extends StatedFragment implements TaskCal
                     BaasUser user = res.value();
                     Log.d("LOG", "The user: " + user);
                     float rankToset;
-                    long l = user.getScope(BaasUser.Scope.REGISTERED).get("feedback");
+                    double l = user.getScope(BaasUser.Scope.REGISTERED).get("feedback");
                     Object o = l;
                     double rank = (double)l;
                     rankToset = getRoundedRank(rank);
